IMS-VT: Add APIs to notify call state changes to VideoPauseTracker.
In some scenarios (like when call is moved to held state), we
may need to clear video pause requests for the connection.
So that subsequent pause request could be correctly handled.
To enable this add APIs to notify call state changes to
VideoPauseTracker and clear the pause requests list when
call moves to hold state.

Change-Id: I1b9d278f0980eec39ea0aff32c3a4b3a450ce34d
CRs-Fixed: 2046005