IMS-VT: Fix video pause issue when both devices are multitasking.
If remote device is in background, the call will be in paused state.
In that state if this device also goes to background, we need to
inform modem that we are currently multi tasking. This is needed so
that when remote device tries to resume video, it should remain
in paused state as long as we are multi tasking.
To enable this, allow pause request even if call is currently in
paused state when the request is received from InCallService.

Change-Id: I0705b7df53e1cfe521ab058a6f5fbe57b61d9008
CRs-Fixed: 2061989